Implement zooming and moving

+++ res/shaders/shader.frag
@@ -1,4 +1,4 @@
-#version 330 core
+#version 400 core
 
 #define MAX_ITERS 100 
 #define PI 3.1415926538
@@ -9,15 +9,15 @@ uniform float zoom;
 
 out vec4 out_Color;
 
-vec2 iter(vec2 prev, vec2 num) {
-    float x = (prev.x * prev.x) - (prev.y * prev.y) + num.x;
-    float y = (2.0 * prev.x * prev.y) + num.y;
+dvec2 iter(dvec2 prev, dvec2 num) {
+    double x = (prev.x * prev.x) - (prev.y * prev.y) + num.x;
+    double y = (2.0 * prev.x * prev.y) + num.y;
 
-    return vec2(x, y);
+    return dvec2(x, y);
 }
 
-int iters_count(vec2 pos) {
-    vec2 prev = vec2(0, 0);
+int iters_count(dvec2 pos) {
+    dvec2 prev = dvec2(0, 0);
 
     for(int i = 0; i < MAX_ITERS; i++) {
         prev = iter(prev, pos);        
@@ -41,7 +41,7 @@ void main()
 
 		vec2 uv = relCoord / screenResolution.x;
     
-    vec2 pos = uv.xy * zoom + centerOffset;
+    dvec2 pos = uv.xy * zoom + centerOffset;
    
     int itersCount = iters_count(pos);
